requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1
requests==2.31.0
